+++ src/models/PaymentIntent.php
@@ -11,6 +11,7 @@ use Craft;
 use craft\commerce\base\GatewayInterface;
 use craft\commerce\base\Model;
 use craft\commerce\elements\Order;
+use craft\commerce\models\Transaction;
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\Plugin as StripePlugin;
 use craft\commerce\stripe\records\Customer as CustomerRecord;
@@ -48,6 +49,11 @@ class PaymentIntent extends Model
     public $orderId;
 
     /**
+     * @var string The Transaction Hash.
+     */
+    public $transactionHash;
+
+    /**
      * @var string Reference
      */
     public $reference;
@@ -78,6 +84,11 @@ class PaymentIntent extends Model
     private $_order;
 
     /**
+     * @var Transaction|null
+     */
+    private $_transaction;
+
+    /**
      * Returns the customer identifier
      *
      * @return string
@@ -88,7 +99,7 @@ class PaymentIntent extends Model
     }
 
     /**
-     * Returns the user element associated with this customer.
+     * Returns the user element associated with this this payment intent.
      *
      * @return User|null
      */
@@ -105,7 +116,7 @@ class PaymentIntent extends Model
     }
 
     /**
-     * Returns the gateway associated with this customer.
+     * Returns the gateway associated with this this payment intent.
      *
      * @return GatewayInterface|null
      */
@@ -119,7 +130,7 @@ class PaymentIntent extends Model
     }
 
     /**
-     * Returns the user element associated with this customer.
+     * Returns the user customer associated with this payment intent.
      *
      * @return Customer|null
      */
@@ -133,7 +144,7 @@ class PaymentIntent extends Model
     }
 
     /**
-     * Returns the gateway associated with this customer.
+     * Returns the gateway associated with this payment intent.
      *
      * @return Order|null
      */
@@ -147,13 +158,27 @@ class PaymentIntent extends Model
     }
 
     /**
+     * Returns the transation associated with this payment intent.
+     *
+     * @return Order|null
+     */
+    public function getTransaction()
+    {
+        if (null === $this->_tranasction) {
+            $this->_transaction = Commerce::getInstance()->getTransactions()->getTransactionByHash($this->transactionHash);
+        }
+
+        return $this->_transaction;
+    }
+
+    /**
      * @inheritdoc
      */
     protected function defineRules(): array
     {
         return [
             [['reference'], 'unique', 'targetAttribute' => ['gatewayId', 'reference'], 'targetClass' => CustomerRecord::class],
-            [['gatewayId', 'customerId', 'reference', 'intentData', 'orderId'], 'required'],
+            [['gatewayId', 'customerId', 'reference', 'intentData', 'orderId', 'transactionHash'], 'required'],
         ];
     }
 }