+++ src/services/PaymentIntents.php
@@ -29,11 +29,21 @@ class PaymentIntents extends Component
      *
      * @return PaymentIntent|null
      */
-    public function getPaymentIntent(int $gatewayId, $orderId, $customerId)
+    public function getPaymentIntent(int $gatewayId, $orderId, $customerId, $transactionHash = null)
     {
-        $result = $this->_createIntentQuery()
-            ->where(['orderId' => $orderId, 'gatewayId' => $gatewayId, 'customerId' => $customerId])
-            ->one();
+        // The `orderId` is not unique across multiple payments (the new partial payment feature) on the same order.
+        // We have added `transactionHash` in commerce-stripe 2.4 to solve this.
+        // TODO make transactionHash required in next major release. Could also drop `orderId`
+        if ($transactionHash) {
+            $result = $this->_createIntentQuery()
+                ->where(['orderId' => $orderId, 'gatewayId' => $gatewayId, 'customerId' => $customerId, 'transactionHash' => $transactionHash])
+                ->one();
+        } else {
+            // This is just in case 3rd party code is calling this method.
+            $result = $this->_createIntentQuery()
+                ->where(['orderId' => $orderId, 'gatewayId' => $gatewayId, 'customerId' => $customerId, 'transactionHash' => null])
+                ->one();
+        }
 
         if ($result !== null) {
             return new PaymentIntent($result);
@@ -86,6 +96,7 @@ class PaymentIntents extends Component
         $record->gatewayId = $paymentIntent->gatewayId;
         $record->customerId = $paymentIntent->customerId;
         $record->orderId = $paymentIntent->orderId;
+        $record->transactionHash = $paymentIntent->transactionHash;
         $record->intentData = $paymentIntent->intentData;
 
         $paymentIntent->validate();
@@ -117,6 +128,7 @@ class PaymentIntents extends Component
                 'customerId',
                 'reference',
                 'orderId',
+                'transactionHash',
                 'intentData',
             ])
             ->from(['{{%stripe_paymentintents}}']);