+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release Notes for Stripe for Craft Commerce
 
+## Unreleased
+
+### Added
+- Added support for the `charge.refund.updated` Stripe webhook event.
+
+### Fixed
+- Fixed a bug where a refund could fail asynchronously.
+
+## 2.3.2.2 - 2021-10-10
+
+### Fixed
+- Fixed a bug that prevented multiple payments on a single order. ([#147](https://github.com/craftcms/commerce-stripe/issues/147))
+
 ## 2.3.2.1 - 2020-11-02
 
 ### Fixed